+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Please add changes at the top. When releasing a version add a new header for tha
 
 ### master
 
+
+ - New edge operation _"Merge parallel edges by attribute"_ makes it possible
+   for the user to merge those parallel edges between two vertices that have the
+   same value for the given edge attribute.
+ - Admins can download the last server log using the link `http://<kite ip>:<kite port>/logs`.
  - When running an EC2 cluster, you cannot directly reference s3 files as before (using
    the format `s3n://AWS_ACCESS_KEY_ID:AWS_SECRET_ACCESS_KEY@bucket/path`), see the changelog
    entry below about the data file prefix notation. Instead, for EC2 cluster we automatically